refactor-constants: same as above for bls381, bn256
