+++ bls381/e12.go
@@ -16,10 +16,6 @@
 
 package bls381
 
-import (
-	"github.com/consensys/gurvy/bls381/fp"
-)
-
 // E12 is a degree-two finite field extension of fp6:
 // C0 + C1w where w^3-v is irrep in fp6
 
@@ -236,189 +232,6 @@ func (z *E12) InverseUnitary(x *E12) *E12 {
 func (z *E12) Conjugate(x *E12) *E12 {
 	z.Set(x)
 	z.C1.Neg(&z.C1)
-	return z
-}
-
-// MulByVW set z to x*(y*v*w) and return z
-// here y*v*w means the E12 element with C1.B1=y and all other components 0
-func (z *E12) MulByVW(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (1,1)
-		var buf E2
-		buf.Set(y)
-		yNR.A1.Add(&buf.A0, &buf.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf.A0)
-	} // end inline: set yNR to (y) * (1,1)
-	result.C0.B0.Mul(&x.C1.B1, &yNR)
-	result.C0.B1.Mul(&x.C1.B2, &yNR)
-	result.C0.B2.Mul(&x.C1.B0, y)
-	result.C1.B0.Mul(&x.C0.B2, &yNR)
-	result.C1.B1.Mul(&x.C0.B0, y)
-	result.C1.B2.Mul(&x.C0.B1, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV set z to x*(y*v) and return z
-// here y*v means the E12 element with C0.B1=y and all other components 0
-func (z *E12) MulByV(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (1,1)
-		var buf E2
-		buf.Set(y)
-		yNR.A1.Add(&buf.A0, &buf.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf.A0)
-	} // end inline: set yNR to (y) * (1,1)
-	result.C0.B0.Mul(&x.C0.B2, &yNR)
-	result.C0.B1.Mul(&x.C0.B0, y)
-	result.C0.B2.Mul(&x.C0.B1, y)
-	result.C1.B0.Mul(&x.C1.B2, &yNR)
-	result.C1.B1.Mul(&x.C1.B0, y)
-	result.C1.B2.Mul(&x.C1.B1, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV2W set z to x*(y*v^2*w) and return z
-// here y*v^2*w means the E12 element with C1.B2=y and all other components 0
-func (z *E12) MulByV2W(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (1,1)
-		var buf E2
-		buf.Set(y)
-		yNR.A1.Add(&buf.A0, &buf.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf.A0)
-	} // end inline: set yNR to (y) * (1,1)
-	result.C0.B0.Mul(&x.C1.B0, &yNR)
-	result.C0.B1.Mul(&x.C1.B1, &yNR)
-	result.C0.B2.Mul(&x.C1.B2, &yNR)
-	result.C1.B0.Mul(&x.C0.B1, &yNR)
-	result.C1.B1.Mul(&x.C0.B2, &yNR)
-	result.C1.B2.Mul(&x.C0.B0, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV2NRInv set z to x*(y*v^2*(1,1)^{-1}) and return z
-// here y*v^2 means the E12 element with C0.B2=y and all other components 0
-func (z *E12) MulByV2NRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (1,1)^{-1}
-		// (yNRInv).A0 = ((y).A0 + (y).A1)/2
-		// (yNRInv).A1 = ((y).A1 - (y).A0)/2
-		buf := *(y)
-		(yNRInv).A0.Add(&buf.A0, &buf.A1)
-		(yNRInv).A1.Sub(&buf.A1, &buf.A0)
-		twoInv := fp.Element{
-			1730508156817200468,
-			9606178027640717313,
-			7150789853162776431,
-			7936136305760253186,
-			15245073033536294050,
-			1728177566264616342,
-		}
-		(yNRInv).A0.MulAssign(&twoInv)
-		(yNRInv).A1.MulAssign(&twoInv)
-	} // end inline: set yNRInv to (y) * (1,1)^{-1}
-
-	result.C0.B0.Mul(&x.C0.B1, y)
-	result.C0.B1.Mul(&x.C0.B2, y)
-	result.C0.B2.Mul(&x.C0.B0, &yNRInv)
-
-	result.C1.B0.Mul(&x.C1.B1, y)
-	result.C1.B1.Mul(&x.C1.B2, y)
-	result.C1.B2.Mul(&x.C1.B0, &yNRInv)
-
-	z.Set(&result)
-	return z
-}
-
-// MulByVWNRInv set z to x*(y*v*w*(1,1)^{-1}) and return z
-// here y*v*w means the E12 element with C1.B1=y and all other components 0
-func (z *E12) MulByVWNRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (1,1)^{-1}
-		// (yNRInv).A0 = ((y).A0 + (y).A1)/2
-		// (yNRInv).A1 = ((y).A1 - (y).A0)/2
-		buf := *(y)
-		(yNRInv).A0.Add(&buf.A0, &buf.A1)
-		(yNRInv).A1.Sub(&buf.A1, &buf.A0)
-		twoInv := fp.Element{
-			1730508156817200468,
-			9606178027640717313,
-			7150789853162776431,
-			7936136305760253186,
-			15245073033536294050,
-			1728177566264616342,
-		}
-		(yNRInv).A0.MulAssign(&twoInv)
-		(yNRInv).A1.MulAssign(&twoInv)
-	} // end inline: set yNRInv to (y) * (1,1)^{-1}
-
-	result.C0.B0.Mul(&x.C1.B1, y)
-	result.C0.B1.Mul(&x.C1.B2, y)
-	result.C0.B2.Mul(&x.C1.B0, &yNRInv)
-
-	result.C1.B0.Mul(&x.C0.B2, y)
-	result.C1.B1.Mul(&x.C0.B0, &yNRInv)
-	result.C1.B2.Mul(&x.C0.B1, &yNRInv)
-
-	z.Set(&result)
-	return z
-}
-
-// MulByWNRInv set z to x*(y*w*(1,1)^{-1}) and return z
-// here y*w means the E12 element with C1.B0=y and all other components 0
-func (z *E12) MulByWNRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (1,1)^{-1}
-		// (yNRInv).A0 = ((y).A0 + (y).A1)/2
-		// (yNRInv).A1 = ((y).A1 - (y).A0)/2
-		buf := *(y)
-		(yNRInv).A0.Add(&buf.A0, &buf.A1)
-		(yNRInv).A1.Sub(&buf.A1, &buf.A0)
-		twoInv := fp.Element{
-			1730508156817200468,
-			9606178027640717313,
-			7150789853162776431,
-			7936136305760253186,
-			15245073033536294050,
-			1728177566264616342,
-		}
-		(yNRInv).A0.MulAssign(&twoInv)
-		(yNRInv).A1.MulAssign(&twoInv)
-	} // end inline: set yNRInv to (y) * (1,1)^{-1}
-
-	result.C0.B0.Mul(&x.C1.B2, y)
-	result.C0.B1.Mul(&x.C1.B0, &yNRInv)
-	result.C0.B2.Mul(&x.C1.B1, &yNRInv)
-
-	result.C1.B0.Mul(&x.C0.B0, &yNRInv)
-	result.C1.B1.Mul(&x.C0.B1, &yNRInv)
-	result.C1.B2.Mul(&x.C0.B2, &yNRInv)
-
-	z.Set(&result)
 	return z
 }
