+++ bls381/e12_test.go
@@ -77,18 +77,6 @@ func TestE12Mul(t *testing.T) {
 	E12check(t, (*E12).Mul, 2)
 }
 
-func TestE12MulByV(t *testing.T) {
-	E12check(t, (*E12).MulByVBinary, 3)
-}
-
-func TestE12MulByVW(t *testing.T) {
-	E12check(t, (*E12).MulByVWBinary, 4)
-}
-
-func TestE12MulByV2W(t *testing.T) {
-	E12check(t, (*E12).MulByV2WBinary, 5)
-}
-
 func TestE12MulByV2NRInv(t *testing.T) {
 	E12check(t, (*E12).MulByV2NRInvBinary, 6)
 }
@@ -154,24 +142,6 @@ func BenchmarkE12Sub(b *testing.B) {
 func BenchmarkE12Mul(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		E12BenchOut.Mul(&E12BenchIn1, &E12BenchIn2)
-	}
-}
-
-func BenchmarkE12MulByV(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		E12BenchOut.MulByVBinary(&E12BenchIn1, &E12BenchIn2)
-	}
-}
-
-func BenchmarkE12MulByVW(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		E12BenchOut.MulByVWBinary(&E12BenchIn1, &E12BenchIn2)
-	}
-}
-
-func BenchmarkE12MulByV2W(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		E12BenchOut.MulByV2WBinary(&E12BenchIn1, &E12BenchIn2)
 	}
 }
 
@@ -291,27 +261,6 @@ func (z *E12) ExptBinary(x, y *E12) *E12 {
 	// if tAbsVal is negative then need to undo the conjugation in order to match the test point
 	z.Conjugate(z) // because tAbsVal is negative
 
-	return z
-}
-
-// MulByVBinary a binary wrapper for MulByV
-func (z *E12) MulByVBinary(x, y *E12) *E12 {
-	yCopy := y.C0.B1
-	z.MulByV(x, &yCopy)
-	return z
-}
-
-// MulByVWBinary a binary wrapper for MulByVW
-func (z *E12) MulByVWBinary(x, y *E12) *E12 {
-	yCopy := y.C1.B1
-	z.MulByVW(x, &yCopy)
-	return z
-}
-
-// MulByV2WBinary a binary wrapper for MulByV2W
-func (z *E12) MulByV2WBinary(x, y *E12) *E12 {
-	yCopy := y.C1.B2
-	z.MulByV2W(x, &yCopy)
 	return z
 }
