+++ bn256/e12.go
@@ -16,10 +16,6 @@
 
 package bn256
 
-import (
-	"github.com/consensys/gurvy/bn256/fp"
-)
-
 // E12 is a degree-two finite field extension of fp6:
 // C0 + C1w where w^3-v is irrep in fp6
 
@@ -248,219 +244,6 @@ func (z *E12) InverseUnitary(x *E12) *E12 {
 func (z *E12) Conjugate(x *E12) *E12 {
 	z.Set(x)
 	z.C1.Neg(&z.C1)
-	return z
-}
-
-// MulByVW set z to x*(y*v*w) and return z
-// here y*v*w means the E12 element with C1.B1=y and all other components 0
-func (z *E12) MulByVW(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (9,1)
-		var buf, buf9 E2
-		buf.Set(y)
-		buf9.Double(&buf).
-			Double(&buf9).
-			Double(&buf9).
-			Add(&buf9, &buf)
-		yNR.A1.Add(&buf.A0, &buf9.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf9.A0)
-	} // end inline: set yNR to (y) * (9,1)
-	result.C0.B0.Mul(&x.C1.B1, &yNR)
-	result.C0.B1.Mul(&x.C1.B2, &yNR)
-	result.C0.B2.Mul(&x.C1.B0, y)
-	result.C1.B0.Mul(&x.C0.B2, &yNR)
-	result.C1.B1.Mul(&x.C0.B0, y)
-	result.C1.B2.Mul(&x.C0.B1, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV set z to x*(y*v) and return z
-// here y*v means the E12 element with C0.B1=y and all other components 0
-func (z *E12) MulByV(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (9,1)
-		var buf, buf9 E2
-		buf.Set(y)
-		buf9.Double(&buf).
-			Double(&buf9).
-			Double(&buf9).
-			Add(&buf9, &buf)
-		yNR.A1.Add(&buf.A0, &buf9.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf9.A0)
-	} // end inline: set yNR to (y) * (9,1)
-	result.C0.B0.Mul(&x.C0.B2, &yNR)
-	result.C0.B1.Mul(&x.C0.B0, y)
-	result.C0.B2.Mul(&x.C0.B1, y)
-	result.C1.B0.Mul(&x.C1.B2, &yNR)
-	result.C1.B1.Mul(&x.C1.B0, y)
-	result.C1.B2.Mul(&x.C1.B1, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV2W set z to x*(y*v^2*w) and return z
-// here y*v^2*w means the E12 element with C1.B2=y and all other components 0
-func (z *E12) MulByV2W(x *E12, y *E2) *E12 {
-	var result E12
-	var yNR E2
-
-	{ // begin inline: set yNR to (y) * (9,1)
-		var buf, buf9 E2
-		buf.Set(y)
-		buf9.Double(&buf).
-			Double(&buf9).
-			Double(&buf9).
-			Add(&buf9, &buf)
-		yNR.A1.Add(&buf.A0, &buf9.A1)
-		{ // begin inline: set &(yNR).A0 to (&buf.A1) * (-1)
-			(&(yNR).A0).Neg(&buf.A1)
-		} // end inline: set &(yNR).A0 to (&buf.A1) * (-1)
-		yNR.A0.AddAssign(&buf9.A0)
-	} // end inline: set yNR to (y) * (9,1)
-	result.C0.B0.Mul(&x.C1.B0, &yNR)
-	result.C0.B1.Mul(&x.C1.B1, &yNR)
-	result.C0.B2.Mul(&x.C1.B2, &yNR)
-	result.C1.B0.Mul(&x.C0.B1, &yNR)
-	result.C1.B1.Mul(&x.C0.B2, &yNR)
-	result.C1.B2.Mul(&x.C0.B0, y)
-	z.Set(&result)
-	return z
-}
-
-// MulByV2NRInv set z to x*(y*v^2*(9,1)^{-1}) and return z
-// here y*v^2 means the E12 element with C0.B2=y and all other components 0
-func (z *E12) MulByV2NRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (9,1)^{-1}
-		// (yNRInv).A0 = (9*(y).A0 + (y).A1)/82
-		// (yNRInv).A1 = (9*(y).A1 - (y).A0)/82
-		copy := *(y)
-
-		var copy9 E2
-		copy9.Double(&copy).
-			Double(&copy9).
-			Double(&copy9).
-			AddAssign(&copy)
-
-		(yNRInv).A0.Add(&copy9.A0, &copy.A1)
-		(yNRInv).A1.Sub(&copy9.A1, &copy.A0)
-
-		buf82inv := fp.Element{
-			15263610803691847034,
-			14617516054323294413,
-			1961223913490700324,
-			3456812345740674661,
-		}
-		(yNRInv).A0.MulAssign(&buf82inv)
-		(yNRInv).A1.MulAssign(&buf82inv)
-	} // end inline: set yNRInv to (y) * (9,1)^{-1}
-
-	result.C0.B0.Mul(&x.C0.B1, y)
-	result.C0.B1.Mul(&x.C0.B2, y)
-	result.C0.B2.Mul(&x.C0.B0, &yNRInv)
-
-	result.C1.B0.Mul(&x.C1.B1, y)
-	result.C1.B1.Mul(&x.C1.B2, y)
-	result.C1.B2.Mul(&x.C1.B0, &yNRInv)
-
-	z.Set(&result)
-	return z
-}
-
-// MulByVWNRInv set z to x*(y*v*w*(9,1)^{-1}) and return z
-// here y*v*w means the E12 element with C1.B1=y and all other components 0
-func (z *E12) MulByVWNRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (9,1)^{-1}
-		// (yNRInv).A0 = (9*(y).A0 + (y).A1)/82
-		// (yNRInv).A1 = (9*(y).A1 - (y).A0)/82
-		copy := *(y)
-
-		var copy9 E2
-		copy9.Double(&copy).
-			Double(&copy9).
-			Double(&copy9).
-			AddAssign(&copy)
-
-		(yNRInv).A0.Add(&copy9.A0, &copy.A1)
-		(yNRInv).A1.Sub(&copy9.A1, &copy.A0)
-
-		buf82inv := fp.Element{
-			15263610803691847034,
-			14617516054323294413,
-			1961223913490700324,
-			3456812345740674661,
-		}
-		(yNRInv).A0.MulAssign(&buf82inv)
-		(yNRInv).A1.MulAssign(&buf82inv)
-	} // end inline: set yNRInv to (y) * (9,1)^{-1}
-
-	result.C0.B0.Mul(&x.C1.B1, y)
-	result.C0.B1.Mul(&x.C1.B2, y)
-	result.C0.B2.Mul(&x.C1.B0, &yNRInv)
-
-	result.C1.B0.Mul(&x.C0.B2, y)
-	result.C1.B1.Mul(&x.C0.B0, &yNRInv)
-	result.C1.B2.Mul(&x.C0.B1, &yNRInv)
-
-	z.Set(&result)
-	return z
-}
-
-// MulByWNRInv set z to x*(y*w*(9,1)^{-1}) and return z
-// here y*w means the E12 element with C1.B0=y and all other components 0
-func (z *E12) MulByWNRInv(x *E12, y *E2) *E12 {
-	var result E12
-	var yNRInv E2
-
-	{ // begin inline: set yNRInv to (y) * (9,1)^{-1}
-		// (yNRInv).A0 = (9*(y).A0 + (y).A1)/82
-		// (yNRInv).A1 = (9*(y).A1 - (y).A0)/82
-		copy := *(y)
-
-		var copy9 E2
-		copy9.Double(&copy).
-			Double(&copy9).
-			Double(&copy9).
-			AddAssign(&copy)
-
-		(yNRInv).A0.Add(&copy9.A0, &copy.A1)
-		(yNRInv).A1.Sub(&copy9.A1, &copy.A0)
-
-		buf82inv := fp.Element{
-			15263610803691847034,
-			14617516054323294413,
-			1961223913490700324,
-			3456812345740674661,
-		}
-		(yNRInv).A0.MulAssign(&buf82inv)
-		(yNRInv).A1.MulAssign(&buf82inv)
-	} // end inline: set yNRInv to (y) * (9,1)^{-1}
-
-	result.C0.B0.Mul(&x.C1.B2, y)
-	result.C0.B1.Mul(&x.C1.B0, &yNRInv)
-	result.C0.B2.Mul(&x.C1.B1, &yNRInv)
-
-	result.C1.B0.Mul(&x.C0.B0, &yNRInv)
-	result.C1.B1.Mul(&x.C0.B1, &yNRInv)
-	result.C1.B2.Mul(&x.C0.B2, &yNRInv)
-
-	z.Set(&result)
 	return z
 }
